backend/requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dict-schema-validator==0.2.0
Flask==1.1.2
Flask-PyMongo==2.3.0
googlemaps==4.4.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
